% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_support.R
\name{browse_checks}
\alias{browse_checks}
\title{developer report on package check process in shiny}
\usage{
browse_checks(con, pnet_obj)
}
\arguments{
\item{con}{SQLiteConnection where tables of check results from rcmdcheck, etc., are stored}

\item{pnet_obj}{list as created by \code{make_pnet_obj}}
}
\description{
developer report on package check process in shiny
}
\note{
dbDisconnect is run if the stop button is pressed to end shiny session.  If
stopped via ctrl-C or error, the connection may remain and require closure.  N.B.  The
source codes for this app are in inst/app1, to simplify contribution to shinyapps.io.
Interactive execution proceeds by copying the server and ui files to temporary folder
where \code{shiny::runApp()} is executed.
}
\examples{
if (interactive()) {
 con = RSQLite::dbConnect(RSQLite::SQLite(), 
  system.file("sqlite/demo9.sqlite", package="BiocBuildTools"), flags=RSQLite::SQLITE_RO)
 print(RSQLite::dbListTables(con))
 load(system.file("pkgnet/demo_pnet2.rda", package="BiocBuildTools"))
 browse_checks(con, demo_pnet2)
}
}
