% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgnet.R
\name{make_pnet_object}
\alias{make_pnet_object}
\title{make a serialized collection of pkgnet reports}
\usage{
make_pnet_object(src_folders, target = "pnet.rds")
}
\arguments{
\item{src_folders}{character() collection of source package folders}

\item{target}{character(1)}
}
\description{
make a serialized collection of pkgnet reports
}
\examples{
if (interactive()) {
 folds = dir(system.file("demo_srcs", package="BiocBuildTools"), full.names=TRUE)
 tf = tempfile()
 make_pnet_object(folds, target=tf)
 x = readRDS(tf)
 x$parody$fun
}
}
